be-id-id-graph-api/requirements_docker.txt
fastapi
uvicorn
pandas
azure-identity
azure-storage-blob
